deps: declare node.js build dependencies with full version string
We leave the "^" in play to support all minor versions of the same major version, hence this does not imply a practical change for builds. But it should trigger dependabot do PRs for each patch version update, triggering asset updates in case. And it makes the dependency graph more explict.

Signed-off-by: MichaIng <micha@dietpi.com>